proper wasm compile with make, pre-copy borders

+++ wasm/compile.sh
@@ -2,21 +2,17 @@
 
 source ../emsdk/emsdk_env.sh --build=Release
 
-jpeg="jpeg-6b"
-cp jconfig.h $jpeg
-list="$jpeg/jutils.c $jpeg/jmemmgr.c $jpeg/jmemnobs.c $jpeg/jcomapi.c $jpeg/jerror.c"
-list="$list $jpeg/jdapimin.c $jpeg/jdcoefct.c $jpeg/jdmarker.c $jpeg/jdhuff.c $jpeg/jdphuff.c $jpeg/jdinput.c $jpeg/jdtrans.c"
-list="$list $jpeg/jcapimin.c $jpeg/jcmaster.c $jpeg/jcmarker.c $jpeg/jchuff.c $jpeg/jcphuff.c $jpeg/jcparam.c $jpeg/jctrans.c"
-
 debug=0
 # --pre-js shell.js
 emflags="-O3 -g0 --closure 1"
 [ $debug == 1 ] && emflags="-O2 -g1 -s ASSERTIONS=1"
 
-emcc -DWASM -I $jpeg ../quantsmooth.c $list -o quantsmooth.html --shell-file shell.html \
-	$emflags -Wno-shift-negative-value \
-	-s EXPORTED_FUNCTIONS="['_malloc', '_free']" -s ALLOW_MEMORY_GROWTH=1
-mv quantsmooth.html index.html
+CFLAGS_LIB="$emflags -Wno-shift-negative-value"
+CFLAGS_APP="-DWASM -DNO_SIMD $emflags --shell-file shell.html -s EXPORTED_FUNCTIONS=\"['_malloc', '_free']\" -s ALLOW_MEMORY_GROWTH=1"
+
+jpeg="jpeg-6b"
+make JPEGSRC="$jpeg" CC="emcc" CFLAGS_LIB="$CFLAGS_LIB" CFLAGS_APP="$CFLAGS_APP" APPNAME="quantsmooth.html" clean all
+test -f quantsmooth.html && mv quantsmooth.html index.html
 
 echo "press enter..."; read dummy < /dev/tty
